refactor :hammer: clases y enums separados de Main.java en sus propios archivos mejorar la legibilidad y estructura del codigo
