+++ src/exceptions/NameOfCityIsBlankException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class NameOfCityIsBlankException extends Exception {
+  public NameOfCityIsBlankException() {
+    super();
+  }
+}
