+++ src/exceptions/NotValidNameOfCityException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class NotValidNameOfCityException extends Exception {
+  public NotValidNameOfCityException() {
+    super();
+  }
+}
